move kernels to green.rs
