+++ bem/src/assembly/dense.rs
@@ -7,6 +7,7 @@ use bempp_traits::bem::{DofMap, FunctionSpace};
 use bempp_traits::cell::ReferenceCellType;
 use bempp_traits::element::FiniteElement;
 use bempp_traits::grid::{Geometry, Grid, Topology};
+use crate::green::{laplace_green, laplace_green_dx, laplace_green_dy};
 
 fn get_quadrature_rule(
     test_celltype: ReferenceCellType,
@@ -104,41 +105,6 @@ fn get_quadrature_rule(
             .unwrap()
         }
     }
-}
-
-fn laplace_green(x: &[f64], y: &[f64], _nx: &[f64], _ny: &[f64]) -> f64 {
-    let inv_dist = 1.0
-        / f64::sqrt(
-            (x[0] - y[0]) * (x[0] - y[0])
-                + (x[1] - y[1]) * (x[1] - y[1])
-                + (x[2] - y[2]) * (x[2] - y[2]),
-        );
-
-    0.25 * std::f64::consts::FRAC_1_PI * inv_dist
-}
-
-fn laplace_green_dx(x: &[f64], y: &[f64], nx: &[f64], _ny: &[f64]) -> f64 {
-    let inv_dist = 1.0
-        / f64::sqrt(
-            (x[0] - y[0]) * (x[0] - y[0])
-                + (x[1] - y[1]) * (x[1] - y[1])
-                + (x[2] - y[2]) * (x[2] - y[2]),
-        );
-    let sum = (y[0] - x[0]) * nx[0] + (y[1] - x[1]) * nx[1] + (y[2] - x[2]) * nx[2];
-
-    0.25 * std::f64::consts::FRAC_1_PI * inv_dist * inv_dist * inv_dist * sum
-}
-
-fn laplace_green_dy(x: &[f64], y: &[f64], _nx: &[f64], ny: &[f64]) -> f64 {
-    let inv_dist = 1.0
-        / f64::sqrt(
-            (x[0] - y[0]) * (x[0] - y[0])
-                + (x[1] - y[1]) * (x[1] - y[1])
-                + (x[2] - y[2]) * (x[2] - y[2]),
-        );
-    let sum = (x[0] - y[0]) * ny[0] + (x[1] - y[1]) * ny[1] + (x[2] - y[2]) * ny[2];
-
-    0.25 * std::f64::consts::FRAC_1_PI * inv_dist * inv_dist * inv_dist * sum
 }
 
 pub fn laplace_single_layer(
