+++ bem/src/lib.rs
@@ -4,3 +4,4 @@
 pub mod assembly;
 pub mod dofmap;
 pub mod function_space;
+pub mod green;
